Add asset loading and logo.png
Copied assets.cmake into the project from PlaylistManager and tweaked CMakeLists.txt to run the file during build. To test, the logo that Laynia's friend drew was thrown in.
Asset loading is needed in order to get the PNG data included in the mod. PlaylistManager has the most recent version of the script.